pyproject.toml
[project]
name = "fastmcp-client-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=3.1.0",
    "python-dotenv>=1.2.2",
    "rich>=14.3.3",
]
